Upgrade generic-pool module
Prior to v2, `generic-pool` accepted callbacks of arbitrary-arity (arbitrarity).
This conflicted with `newrelic` wrapping callbacks during its instrumentation,
which appeared to change the arity on them. As a result, the DB connection would
come back in the `error` argument and we'd never be able to do anything.

The rest of the changes in `generic-pool` v2 appear minor, and @temas
successfully tested them on a separate branch not too long ago. The pool is
only used in one place in `dal`, which still works just fine.