Fix API explorer showing invalid Twitter endpoints
Since it pulls from the synclet names, the explorer was showing `sent_messages`
and `received_messages`, but those aren't endpoints. First, we filter those out
with the `FILTER` in `resources.js` to hide them, then we add `direct_messages`
as an `aka` on the two synclets so it'll show up in the list. We want the `aka`
on both for consistency, but also so that `taskman#fresh` deals with both
properly. Then since it's on both, we need to dedup the endpoint list that
`resources` reports.