Make photos synclet behave when there are 0 albums
Instead of looping forever, the photos synclet can now tell whether the albums
synclet has run at all. If it hasn't, we'll wait, otherwise there's nothing do
to.

If there's a problem with the albums synclet and the run time doesn't get saved,
it's possible for photos to loop forever. We'll figure that out when we hit it.